requirements.txt
pymongo >=4.3.3
python-dotenv >= 1.0.0
beautifulsoup4 >= 4.11.2
requests
lxml
